pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "fastapi-template"
version = "0.1.0"
description = "A modern FastAPI backend starter template."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "synergy" }]
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "sqlalchemy[asyncio]>=2.0.40,<3.0.0",
  "alembic>=1.16.0,<2.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "aiosqlite>=0.21.0,<1.0.0",
  "structlog>=25.3.0,<26.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "PyJWT>=2.9.0,<3.0.0",
  "bcrypt>=4.2.0,<5.0.0",
  "redis[asyncio]>=5.0.0,<6.0.0",
  "arq>=0.26.0,<1.0.0",
  "slowapi>=0.1.9,<1.0.0",
  "sentry-sdk[fastapi]>=2.0.0,<3.0.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.26.0,<1.0.0",
  "ruff>=0.11.0,<1.0.0",
  "mypy>=1.15.0,<2.0.0",
  "pre-commit>=4.2.0,<5.0.0",
  "pytest-cov>=7.1.0",
  "asgi-lifespan>=2.1.0,<3.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC", "RUF"]
ignore = ["B008", "RUF001", "RUF002", "RUF003"]  # RUF001/002/003: 中文全角标点属于正常用法

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
# 所有 async test/fixture 共用同一 event loop，
# 避免模块级 SQLAlchemy engine 和 Redis pool 跨 loop 导致的 "Event loop is closed"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
branch = true   # measure branch coverage in addition to line coverage
# 现已由集成测试覆盖：cache.py / redis.py / arq.py / worker.py 的任务函数
omit = [
    "app/core/sentry.py",    # 需要真实 Sentry DSN 才能覆盖
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
show_missing = true

[tool.mypy]
python_version = "3.12"
plugins = []
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true
pretty = true
exclude = ["alembic"]

